move the validation and plan tests into dedicated subpackages
