+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/api/scala/batch/table/stringexpr/CalcStringExpressionTest.scala
@@ -16,19 +16,19 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.api.scala.batch.table
+package org.apache.flink.table.api.scala.batch.table.stringexpr
 
 import java.sql.{Date, Time, Timestamp}
 
 import org.apache.flink.api.scala._
-import org.apache.flink.table.api.scala._
 import org.apache.flink.api.scala.util.CollectionDataSets
-import org.apache.flink.table.api.{TableEnvironment, Types}
+import org.apache.flink.table.api.scala._
 import org.apache.flink.table.api.scala.batch.utils.LogicalPlanFormatUtils
+import org.apache.flink.table.api.{TableEnvironment, Types}
 import org.apache.flink.table.expressions.Literal
 import org.junit._
 
-class CalcPlanTest {
+class CalcStringExpressionTest {
 
   @Test
   def testSimpleSelectAllWithAs(): Unit = {
